Testing

+++ c_lib/Sandworm.h
@@ -34,6 +34,6 @@ void Initialize_Sandworm( Sandworm_t* p_sw, float Lin_pos, float Rot_pos, float 
  * @param Lin_Vel The desired linear actuator velocity
  * @param Rot_vel The desired Rotary actuator velocity
  */
-void Sandworm_Speed( Sandworm_t* p_sw, uint16_t Lin_Vel, int Rot_vel );
+void Sandworm_Speed( Sandworm_t* p_sw, int Lin_Vel, int Rot_vel );
 
 #endif
